% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_geary.R
\name{ww_global_geary_c}
\alias{ww_global_geary_c}
\alias{ww_global_geary_c_vec}
\alias{ww_global_geary_pvalue}
\alias{ww_global_geary_pvalue_vec}
\alias{ww_global_geary}
\title{Global Moran's I statistic}
\usage{
ww_global_geary_c(data, ...)

ww_global_geary_c_vec(
  truth,
  estimate,
  nb,
  wt,
  allow_zero = TRUE,
  na_rm = TRUE,
  ...
)

ww_global_geary_pvalue(data, ...)

ww_global_geary_pvalue_vec(
  truth,
  estimate,
  nb,
  wt,
  alternative = "greater",
  nsim = 499,
  allow_zero = NULL,
  na_rm = TRUE,
  ...
)

ww_global_geary(
  data,
  truth,
  estimate,
  nb,
  wt,
  alternative = "greater",
  nsim = 499,
  allow_zero = NULL,
  na_rm = TRUE,
  ...
)
}
\arguments{
\item{data}{A \code{data.frame} containing the columns specified by the \code{truth}
and \code{estimate} arguments.}

\item{...}{Not currently used.}

\item{truth}{The column identifier for the true results
(that is \code{numeric}). This should be an unquoted column name although
this argument is passed by expression and supports
\link[rlang:topic-inject]{quasiquotation} (you can unquote column
names). For \verb{_vec()} functions, a \code{numeric} vector.}

\item{estimate}{The column identifier for the predicted
results (that is also \code{numeric}). As with \code{truth} this can be
specified different ways but the primary method is to use an
unquoted variable name. For \verb{_vec()} functions, a \code{numeric} vector.}

\item{nb}{a neighbor list object for example as created by \code{st_contiguity()}.}

\item{wt}{a weights list as created by \code{st_weights()}.}

\item{allow_zero}{If \code{TRUE}, assigns zero as lagged value to zone without neighbors.}

\item{na_rm}{A \code{logical} value indicating whether \code{NA}
values should be stripped before the computation proceeds.}

\item{alternative}{default \code{"two.sided"}. Should be one of \code{"greater"}, \code{"less"}, or \code{"two.sided"} to specify the alternative hypothesis.}

\item{nsim}{number of simulations to run.}
}
\value{
A tibble with columns .metric, .estimator, and .estimate.
For grouped data frames, the number of rows returned will be the same as the number of groups times the number of metrics.
For ww_global_geary_c_vec(), a single numeric value (or NA).
}
\description{
Calculate the global Moran's I statistic for model residuals.
\code{ww_global_geary_c()} returns the statistic itself, while
\code{ww_global_geary_pvalue()} returns the associated p value.
\code{ww_global_geary()} returns both.
}
\examples{
data(guerry, package = "sfdep")

guerry_modeled <- guerry
guerry_lm <- lm(crime_pers ~ literacy, guerry_modeled)
guerry_modeled$predictions <- predict(guerry_lm, guerry_modeled)

ctg <- st_contiguity(guerry)
wts <- st_weights(ctg)

ww_global_geary_c(guerry_modeled, crime_pers, predictions, ctg, wts)
ww_global_geary(guerry_modeled, crime_pers, predictions, ctg, wts)

}
