% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{gmfr_rust}
\alias{gmfr_rust}
\title{Calculate geometric mean functional relationship parameters}
\usage{
gmfr_rust(truth, estimate, corsign)
}
\description{
Calculate geometric mean functional relationship parameters
}
