% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_moran.R
\name{ww_local_moran_i}
\alias{ww_local_moran_i}
\alias{ww_local_moran_i_vec}
\alias{ww_local_moran_pvalue}
\alias{ww_local_moran_pvalue_vec}
\alias{ww_local_moran}
\title{Local Moran's I statistic}
\usage{
ww_local_moran_i(data, ...)

ww_local_moran_i_vec(
  truth,
  estimate,
  nb,
  wt,
  alternative = "two.sided",
  nsim = 499,
  na_rm = TRUE,
  seed = .Random.seed,
  ...
)

ww_local_moran_pvalue(data, ...)

ww_local_moran_pvalue_vec(
  truth,
  estimate,
  nb,
  wt,
  alternative = "two.sided",
  nsim = 499,
  na_rm = TRUE,
  seed = .Random.seed,
  ...
)

ww_local_moran(
  data,
  truth,
  estimate,
  nb,
  wt,
  alternative = "greater",
  nsim = 499,
  na_rm = TRUE,
  seed = .Random.seed,
  ...
)
}
\arguments{
\item{data}{A \code{data.frame} containing the columns specified by the \code{truth}
and \code{estimate} arguments.}

\item{...}{Not currently used.}

\item{truth}{The column identifier for the true results
(that is \code{numeric}). This should be an unquoted column name although
this argument is passed by expression and supports
\link[rlang:topic-inject]{quasiquotation} (you can unquote column
names). For \verb{_vec()} functions, a \code{numeric} vector.}

\item{estimate}{The column identifier for the predicted
results (that is also \code{numeric}). As with \code{truth} this can be
specified different ways but the primary method is to use an
unquoted variable name. For \verb{_vec()} functions, a \code{numeric} vector.}

\item{nb}{a neighbor list object for example as created by \code{st_contiguity()}.}

\item{wt}{a weights list as created by \code{st_weights()}.}

\item{alternative}{default \code{"two.sided"}. Should be one of \code{"greater"}, \code{"less"}, or \code{"two.sided"} to specify the alternative hypothesis.}

\item{nsim}{The number of simulations to run.}

\item{na_rm}{A \code{logical} value indicating whether \code{NA}
values should be stripped before the computation proceeds.}

\item{seed}{A random seed to use for all metric calculations.
Defaults to the current seed.}
}
\value{
A tibble with columns .metric, .estimator, and .estimate and `nrow(data)` rows of values.
For grouped data frames, the number of rows returned will be the same as the number of groups.
For ww_local_moran_i_vec(), a numeric vector of `length(truth)` (or NA).
}
\description{
Calculate the local Moran's I statistic for model residuals.
}
