% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ny_trees}
\alias{ny_trees}
\title{Number of trees and aboveground biomass for Forest Inventory and Analysis plots in New York State}
\format{
An sf object using EPSG 5070: NAD83 / Conus Albers (in meters), with 5,303 rows and 5 columns:
\describe{
\item{yr}{The year measurements were taken.}
\item{plot}{A unique identifier signifying the plot measurements were taken at.}
\item{n_trees}{The number of trees present on a plot.}
\item{agb}{The total aboveground biomass at the plot location, in pounds.}
\item{geometry}{The centroid of the plot location.}
}
}
\source{
Tidymodels team, "Multi-scale Assessment". Retrieved 30 November 2022. Available at \url{https://www.tidymodels.org/learn/work/multi-scale/}
}
\usage{
ny_trees
}
\description{
This data is adapted from the Tidymodels "multi-scale assessment" lesson, available at \url{https://www.tidymodels.org/learn/work/multi-scale/}.
The original data is derived from the Forest Inventory and Analysis program,
implemented by the US Department of Agriculture's Forest Service.
}
\keyword{datasets}
