% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_of_applicability.R
\name{predict.ww_area_of_applicability}
\alias{predict.ww_area_of_applicability}
\title{Predict from a `ww_area_of_applicability`}
\usage{
\method{predict}{ww_area_of_applicability}(object, new_data, ...)
}
\arguments{
\item{object}{A `ww_area_of_applicability` object.}

\item{new_data}{A data frame or matrix of new samples.}

\item{...}{Not used.}
}
\value{
A tibble of predictions, with two columns: `di`, numeric, contains the
"dissimilarity index" of each point in `new_data`, while `aoa`, logical,
contains whether a row is inside (`TRUE`) or outside (`FALSE`) the area of
applicability.

Note that this function is often called using
[terra::predict()], in which case `aoa` will be converted to numeric
implicitly; `1` values correspond to cells "inside" the area of applicability
and `0` corresponds to cells "outside" the AOA.

The number of rows in the tibble is guaranteed
to be the same as the number of rows in `new_data`. Rows with `NA` predictor
values will have `NA` `di` and `aoa` values.
}
\description{
Predict from a `ww_area_of_applicability`
}
\details{
The function computes the distance indices of the new data and
whether or not they are "inside" the area of applicability.
}
\examples{
\dontshow{if (rlang::is_installed("vip")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(vip)
train <- gen_friedman(1000, seed = 101)  # ?vip::gen_friedman
test <- train[701:1000, ]
train <- train[1:700, ]
pp <- stats::ppr(y ~ ., data = train, nterms = 11)
importance <- vi_permute(
  pp,
  target = "y",
  metric = "rsquared",
  pred_wrapper = predict
)

aoa <- ww_area_of_applicability(y ~ ., train, test, importance = importance)
predict(aoa, test)
\dontshow{\}) # examplesIf}
}
\seealso{
Other area of applicability functions: 
\code{\link{ww_area_of_applicability}()}
}
\concept{area of applicability functions}
