% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waywiser-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{st_contiguity}
\alias{st_dist_band}
\alias{st_knn}
\alias{st_weights}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{sfdep}{\code{\link[sfdep]{st_contiguity}}, \code{\link[sfdep]{st_dist_band}}, \code{\link[sfdep]{st_knn}}, \code{\link[sfdep]{st_weights}}}
}}

