% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{spds_rust}
\alias{spds_rust}
\title{Return the systematic sum product-difference from Ji and Gallo (2006)}
\usage{
spds_rust(truth, estimate, corsign)
}
\description{
Return the systematic sum product-difference from Ji and Gallo (2006)
}
