% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{spdu_rust}
\alias{spdu_rust}
\title{Return the unsystematic sum product-difference from Ji and Gallo (2006)}
\usage{
spdu_rust(truth, estimate, corsign)
}
\description{
Return the unsystematic sum product-difference from Ji and Gallo (2006)
}
