% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_imports.R
\docType{package}
\name{waywiser-package}
\alias{waywiser}
\alias{waywiser-package}
\title{waywiser: Yardstick Extensions for Measuring Spatial Structure in Model Residuals}
\description{
Predictive models of spatial data may have spatially structured errors, with "hot spots" of higher than expected error clustered geographically due to spatial structure in the underlying data. These functions provide methods for measuring the spatial structure of model errors for models, and are particularly useful for models fit using the 'tidymodels' framework.
}
\author{
\strong{Maintainer}: Michael Mahoney \email{mike.mahoney.218@gmail.com} (\href{https://orcid.org/0000-0003-2402-304X}{ORCID})

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
