% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{worldclim_simulation}
\alias{worldclim_simulation}
\title{Simulated data based on WorldClim Bioclimatic variables}
\format{
An sf object with 10,000 rows and 6 columns:
\describe{
\item{bio2}{Mean Diurnal Range (Mean of monthly (max temp - min temp))}
\item{bio10}{Mean Temperature of Warmest Quarter}
\item{bio13}{Precipitation of Wettest Month}
\item{bio19}{Precipitation of Coldest Quarter}
\item{geometry}{The location of the sampled point.}
\item{response}{A virtual species distribution, generated using the \code{generateSpFromPCA()} function from the \code{virtualspecies} package.}
}
}
\source{
\url{https://www.worldclim.org}
}
\usage{
worldclim_simulation
}
\description{
This data is adapted from the CAST vignette
\code{vignette("cast02-AOA-tutorial", package = "CAST")}.
The original data is derived from the Worldclim global climate variables.
}
\keyword{datasets}
