% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{ww_make_polygon_neighbors}
\alias{ww_make_polygon_neighbors}
\title{Make 'nb' objects from polygon geometries}
\usage{
ww_make_polygon_neighbors(data, ...)
}
\arguments{
\item{data}{An \code{sfc_POLYGON} or \code{sfc_MULTIPOLYGON} object.}

\item{...}{Additional arguments passed to \code{\link[spdep:poly2nb]{spdep::poly2nb()}}.}
}
\value{
An object of class "nb"
}
\description{
This function is an extremely thin wrapper around \code{\link[spdep:poly2nb]{spdep::poly2nb()}},
renamed to use the waywiser "ww" prefix.
}
\details{
These functions can be used for geographic or projected coordinate reference
systems and expect 2D data.
}
\examples{
ww_make_polygon_neighbors(guerry)

}
